+++ src/Elements/Geometry/Contour.cs
@@ -24,9 +24,15 @@ namespace Elements.Geometry
                 var a = curves[i];
                 var next = i == curves.Count - 1 ? 0 : i + 1;
                 var b = curves[next];
-                if (!a.PointAt(1).IsAlmostEqualTo(b.PointAt(0)))
+                if (a.PointAt(1).IsAlmostEqualTo(b.PointAt(0)) || 
+                    a.PointAt(1).IsAlmostEqualTo(b.PointAt(1)) ||
+                    a.PointAt(0).IsAlmostEqualTo(b.PointAt(0)))
                 {
-                    throw new ArgumentException($"The contour could not be constructed. The end of curve {i} and the start of curve {next} are not continguous.");
+                    continue;
+                }
+                else
+                {
+                    throw new ArgumentException($"The contour could not be constructed. Curves {i} and {next} are not continguous.");
                 }
             }
             this._curves = curves;
@@ -35,7 +41,6 @@ namespace Elements.Geometry
         /// <summary>
         /// Get the enumerator for the collection of curves.
         /// </summary>
-        /// <returns></returns>
         public IEnumerator<Curve> GetEnumerator()
         {
             return _curves.GetEnumerator();
@@ -51,10 +56,33 @@ namespace Elements.Geometry
             var verts = new List<Vector3>();
             foreach(var c in this._curves)
             {
-                verts.AddRange(c.RenderVertices());
+                var v = c.RenderVertices();
+                if(verts.Any())
+                {
+                    if(verts.Last().IsAlmostEqualTo(v.First()))
+                    {
+                        verts.AddRange(v.Skip(1));
+                    }
+                    else if(verts.Last().IsAlmostEqualTo(v.Last()))
+                    {
+                        var revVerts = v.Reverse();
+                        verts.AddRange(revVerts.Skip(1));
+                    }
+                    else
+                    {
+                        verts.AddRange(v);  
+                    }
+                }
+                else
+                {
+                    // We skip the first vert assuming that the polygon
+                    // is closed and that the final set of vertices
+                    // will contain the "last" vert.
+                    verts.AddRange(v.Skip(1));
+                }
             }
 
-            return new Polygon(verts.Distinct().ToList());
+            return new Polygon(verts);
         }
 
         IEnumerator IEnumerable.GetEnumerator()