+++ src/Elements/Geometry/Line.cs
@@ -190,7 +190,7 @@ namespace Elements.Geometry
         }
 
         /// <summary>
-        /// Divide the line into as many segments of length l as possible.
+        /// Divide the line into as many segments of the provided length as possible.
         /// </summary>
         /// <param name="l">The length.</param>
         /// <param name="removeShortSegments">A flag indicating whether segments shorter than l should be removed.</param>
@@ -219,6 +219,49 @@ namespace Elements.Geometry
             }
             return lines;
         }
+        
+
+        /// <summary>
+        /// Divide the line into as many segments of the provided length as possible.
+        /// Divisions will be centered along the line.
+        /// </summary>
+        /// <param name="l">The length.</param>
+        /// <returns></returns>
+        public List<Line> DivideByLengthFromCenter(double l)
+        {
+            var lines = new List<Line>();
+
+            var localLength = this.Length();
+            if(localLength <= l)
+            {
+                lines.Add(this);
+                return lines; 
+            }
+
+            var divs = (int)(localLength/l);
+            var span = divs * l;
+            var halfSpan = span/2;
+            var mid = this.PointAt(0.5);
+            var dir = this.Direction();
+            var start = mid - dir * halfSpan;
+            var end = mid + dir * halfSpan;
+            if(!this.Start.IsAlmostEqualTo(start))
+            {
+                lines.Add(new Line(this.Start, start));
+            }
+            for(var i=0; i<divs; i++)
+            {
+                var p1 = start + (i * l) * dir;
+                var p2 = p1 + dir * l;
+                lines.Add(new Line(p1, p2));
+            }
+            if(!this.End.IsAlmostEqualTo(end))
+            {
+                lines.Add(new Line(end, this.End));
+            }
+        
+            return lines;
+        }
 
         /// <summary>
         /// Divide the line into n+1 equal segments.
@@ -239,6 +282,151 @@ namespace Elements.Geometry
                 lines.Add(new Line(a, b));
             }
             return lines;
+        }
+
+        /// <summary>
+        /// Offset the line. The offset direction will be defined by 
+        /// Direction X Vector3.ZAxis.
+        /// </summary>
+        /// <param name="distance">The distance to offset.</param>
+        /// <param name="flip">Flip the offset direction.</param>
+        /// <returns></returns>
+        public Line Offset(double distance, bool flip)
+        {
+            var offsetVector = this.Direction().Cross(Vector3.ZAxis);
+            if (flip)
+            {
+                offsetVector = offsetVector.Negate();
+            }
+            var a = Start + offsetVector * distance;
+            var b = End + offsetVector * distance;
+            return new Line(a, b);
+        }
+
+        /// <summary>
+        /// Trim this line to the trimming curve.
+        /// </summary>
+        /// <param name="line">The curve to which to trim.</param>
+        /// <param name="flip">Should the trim direction be reversed?</param>
+        /// <returns>A new line, or null if this line does not intersect the trimming line.</returns>
+        public Line TrimTo(Line line, bool flip = false)
+        {
+            var d1 = this.Direction();
+            var d2 = line.Direction();
+
+            var n = d2.Cross(Vector3.ZAxis);
+            var p = new Plane(line.Start, n);
+
+            if (this.Intersects(p, out Vector3 result))
+            {
+                if (flip)
+                {
+                    return new Line(result, this.End);
+                }
+                else
+                {
+                    return new Line(this.Start, result);
+                }
+            }
+
+            return null;
+        }
+        
+        /// <summary>
+        /// Extend this line to the trimming curve.
+        /// </summary>
+        /// <param name="line">The curve to which to extend.</param>
+        /// <returns>A new line, or null if this line does not intersect the trimming line.</returns>
+        public Line ExtendTo(Line line)
+        {
+            var d1 = this.Direction();
+            var d2 = line.Direction();
+            
+            // Extend the line and trim in one direction.
+            var d3 = this.Start + d1 * 1000000;
+            var temp = new Line(this.Start, d3);
+            var trim = temp.TrimTo(line);
+            if(trim != null)
+            {
+                return trim;
+            }
+
+            // Extend the line and trim in the other direction.
+            d3 = this.Start - d1 * 1000000;
+            temp = new Line(this.End, d3);
+            trim = temp.TrimTo(line);
+            return trim;
+        }
+
+        /// <summary>
+        /// Create a fillet arc between this line and the target. 
+        /// </summary>
+        /// <param name="target">The line with which to fillet.</param>
+        /// <param name="radius">The radius of the fillet.</param>
+        /// <returns>An arc, or null if no fillet can be calculated.</returns>
+        public Arc Fillet(Line target, double radius)
+        {
+            var d1 = this.Direction();
+            var d2 = target.Direction();
+            if (d1.IsParallelTo(d2))
+            {
+                throw new Exception("The fillet could not be created. The lines are parallel");
+            }
+            
+            var r1 = new Ray(this.Start, d1);
+            var r2 = new Ray(target.Start, d2);
+            if(!r1.Intersects(r2, out Vector3 result, true))
+            {
+                return null;
+            }
+
+            // Construct new vectors that both
+            // point away from the projected intersection
+            var newD1 = (this.PointAt(0.5) - result).Normalized();
+            var newD2 = (target.PointAt(0.5) - result).Normalized();
+
+            var theta = newD1.AngleTo(newD2) * Math.PI / 180.0;
+            var halfTheta = theta/2.0;
+            var h = radius / Math.Sin(halfTheta);
+            var centerVec = newD1.Average(newD2).Normalized();
+            var arcCenter = result + centerVec * h;
+            
+            // Find the closest points from the arc
+            // center to the adjacent curves.
+            var p1 = arcCenter.ClosestPointOn(this);
+            var p2 = arcCenter.ClosestPointOn(target);
+
+            // Find the angle of both segments relative to the fillet arc.
+            // ATan2 assumes the origin, so correct the coordinates
+            // by the offset of the center of the arc.
+            var angle1 = Math.Atan2(p1.Y - arcCenter.Y, p1.X - arcCenter.X) * 180.0/Math.PI;
+            var angle2 = Math.Atan2(p2.Y - arcCenter.Y, p2.X - arcCenter.X) * 180.0/Math.PI;
+            
+            // ATan2 will provide negative angles in the "lower" quadrants
+            // Ensure that these values are 180d -> 360d
+            angle1 = (angle1 + 360) % 360;
+            angle2 = (angle2 + 360) % 360;
+            angle2 = angle2 == 0.0 ? 360.0 : angle2;
+
+            // We only support CCW wound arcs. 
+            // For arcs that with start angles <1d, convert
+            // the arc back to a negative value.
+            var arc = new Arc(arcCenter,
+                           radius,
+                           angle1 > angle2 ? angle1 - 360.0 : angle1,
+                           angle2);
+
+            // Get the complimentary arc and choose
+            // the shorter of the two arcs.
+            var complement = arc.Complement();
+            if(arc.Length() < complement.Length())
+            {
+                return arc;
+            }
+            else
+            {
+                return complement;
+            }
         }
 
         /// <summary>